app/requirements.txt
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
pydantic==2.5.2
tenacity==8.2.2
prometheus-client==0.19.0
python-json-logger==2.0.7
flask==3.0.0
flask-cors==4.0.0
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
pytest-flask==1.3.0